requirements.txt
streamlit
pandas
numpy
pymupdf
faiss-cpu
openai
python-dotenv
tqdm
requests
